* Bundles/FLAC/FLAC.h           make it compile with non-fragile ABI         * Bundles/Musepack/Musepack.m           fix potentially uninitialized return value         * CynthiuneFadingTextField.h           float -> CGFloat transition         * CynthiunePauseButton.m         * MBResultsPanel.m         * PlaylistViewController.m         * Player.m           NS(U)Integer transitions, and fixing some NSLog           format strings           All this together makes the Player fading when start playing           a song work again on amd64
git-svn-id: svn+ssh://svn.savannah.nongnu.org/gap/trunk@2218 d25b4930-3292-e311-b074-0010c6de5da8

+++ user-apps/Cynthiune/CynthiuneFadingTextField.h
@@ -38,10 +38,10 @@
   unsigned int numberOfIterations;
   unsigned int iteration;
 
-  float red;
-  float green;
-  float blue;
-  float alpha;
+  CGFloat red;
+  CGFloat green;
+  CGFloat blue;
+  CGFloat alpha;
 }
 
 - (void) setStringValue: (NSString *) string;
